refactor: move dtos to separate packages
