+++ passenger/src/main/java/com/modsen/passenger/controller/openapi/RatingControllerOpenApi.java
@@ -1,10 +1,10 @@
 package com.modsen.passenger.controller.openapi;
 
-import com.modsen.passenger.dto.ErrorResponse;
-import com.modsen.passenger.dto.PassengerResponse;
-import com.modsen.passenger.dto.RatingResponse;
-import com.modsen.passenger.dto.ScoreRequest;
-import com.modsen.passenger.dto.ValidationErrorResponse;
+import com.modsen.passenger.dto.response.ErrorResponse;
+import com.modsen.passenger.dto.response.PassengerResponse;
+import com.modsen.passenger.dto.response.RatingResponse;
+import com.modsen.passenger.dto.request.ScoreRequest;
+import com.modsen.passenger.dto.response.ValidationErrorResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.ExampleObject;
