+++ passenger/src/main/java/com/modsen/passenger/dto/response/ErrorResponse.java
@@ -1,16 +1,14 @@
-package com.modsen.passenger.dto;
+package com.modsen.passenger.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.List;
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class ValidationErrorResponse {
+public class ErrorResponse {
 
     private Integer statusCode;
-    private List<String> errors;
+    private String errorMessage;
 }
