+++ passenger/src/main/java/com/modsen/passenger/dto/response/ValidationErrorResponse.java
@@ -1,14 +1,16 @@
-package com.modsen.passenger.dto;
+package com.modsen.passenger.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.util.List;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class ErrorResponse {
+public class ValidationErrorResponse {
 
     private Integer statusCode;
-    private String errorMessage;
+    private List<String> errors;
 }
