+++ passenger/src/main/java/com/modsen/passenger/mapper/PassengerMapper.java
@@ -1,7 +1,7 @@
 package com.modsen.passenger.mapper;
 
-import com.modsen.passenger.dto.PassengerCreate;
-import com.modsen.passenger.dto.PassengerResponse;
+import com.modsen.passenger.dto.request.PassengerCreate;
+import com.modsen.passenger.dto.response.PassengerResponse;
 import com.modsen.passenger.model.Passenger;
 import org.mapstruct.Mapper;
 import org.mapstruct.ReportingPolicy;
