+++ passenger/src/main/java/com/modsen/passenger/service/PassengerService.java
@@ -1,8 +1,8 @@
 package com.modsen.passenger.service;
 
-import com.modsen.passenger.dto.PassengerCreate;
-import com.modsen.passenger.dto.PassengerResponse;
-import com.modsen.passenger.dto.PassengerUpdate;
+import com.modsen.passenger.dto.request.PassengerCreate;
+import com.modsen.passenger.dto.response.PassengerResponse;
+import com.modsen.passenger.dto.request.PassengerUpdate;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
