+++ passenger/src/main/java/com/modsen/passenger/service/RatingService.java
@@ -1,6 +1,6 @@
 package com.modsen.passenger.service;
 
-import com.modsen.passenger.dto.RatingResponse;
+import com.modsen.passenger.dto.response.RatingResponse;
 
 public interface RatingService {
