+++ passenger/src/main/java/com/modsen/passenger/service/impl/RatingServiceImpl.java
@@ -1,6 +1,6 @@
 package com.modsen.passenger.service.impl;
 
-import com.modsen.passenger.dto.RatingResponse;
+import com.modsen.passenger.dto.response.RatingResponse;
 import com.modsen.passenger.exception.PassengerNotFoundException;
 import com.modsen.passenger.exception.RatingNotFoundException;
 import com.modsen.passenger.mapper.RatingMapper;
